+++ src/public/common.ts
@@ -15,6 +15,7 @@ import '@angular/core';
 import '@angular/common';
 import '@angular/http';
 import '@angular/router';
+import "lodash";
 // RxJS
 import 'rxjs';
 // Other vendors for example jQuery, Lodash or Bootstrap