+++ src/public/components/app-layout/applayout.component.ts
@@ -1,10 +1,11 @@
 import { Component } from '@angular/core';
 import { ROUTER_DIRECTIVES }  from '@angular/router';
+import "./applayout.component.less";
 
 @Component({
   selector: 'app-layout',
   template: require('./applayout.component.html'),
-  styles: [require("./applayout.component.less")],
+  styleUrls: [("./applayout.component.less")],
   directives: [ROUTER_DIRECTIVES]
 })