+++ src/public/components/data-table/datatable.component.ts
@@ -3,7 +3,7 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'data-table',
   template: require('./datatable.component.html'),
-  styles: [require("./datatable.component.less")]
+  //styles: [require("./datatable.component.less")]
 })
 
 export class DataTableComponent { }