+++ src/public/routes/demo1/demo1.component.ts
@@ -4,7 +4,7 @@ import { ContextLayoutService } from "../../services/contexttitle.service";
 @Component({
   selector: 'demo1',
   template: require('./demo1.component.html'),
-  styles: [require("./demo1.component.less")]
+  //styles: [require("./demo1.component.less")]
 })
 
 export class Demo1Component {